Switch to long

+++ src/Main.java
@@ -1,9 +1,9 @@
 public class Main {
 
     public static void main(String[] args) {
-        int[] sales = new int[] {1, 2, 300, 200};
+        long[] sales = new long[] {1, 2, 300, 200};
         SalesManager salesManger = new SalesManager(sales);
-int max = salesManger.max();
+long max = salesManger.max();
 System.out.println(max);
     }
 }
